// frontend/Angular-workflow-diagram2-main/src/app/api/model/solicitudResponse.ts
/**
 * Workflow Departamental API
 *
 * Contact: soporte@workflow.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventoHistorialResponse } from './eventoHistorialResponse';


export interface SolicitudResponse { 
    id?: string;
    codigoSeguimiento?: string;
    titulo?: string;
    descripcion?: string;
    prioridad?: SolicitudResponse.PrioridadEnum;
    estado?: SolicitudResponse.EstadoEnum;
    departamentoActual?: string;
    usuarioCreador?: string;
    usuarioAsignado?: string;
    workflowDefinitionId?: string;
    tareaActualId?: string;
    tareaActualNombre?: string;
    historial?: Array<EventoHistorialResponse>;
    fechaCreacion?: string;
    fechaActualizacion?: string;
    fechaLimiteAtencion?: string;
    fechaPrimeraAlertaSla?: string;
    fechaEscalamientoSla?: string;
    archivosAdjuntos?: Array<{id?: string; nombreOriginal?: string; tipoContenido?: string; tamanoBytes?: number; subidoPor?: string; fechaSubida?: string}>;
    datosFormulario?: { [key: string]: any; };
    estadoSla?: SolicitudResponse.EstadoSlaEnum;
    minutosRestantesSla?: number;
    totalEventos?: number;
}
export namespace SolicitudResponse {
    export const PrioridadEnum = {
        Baja: 'BAJA',
        Media: 'MEDIA',
        Alta: 'ALTA',
        Urgente: 'URGENTE'
    } as const;
    export type PrioridadEnum = typeof PrioridadEnum[keyof typeof PrioridadEnum];
    export const EstadoEnum = {
        Pendiente: 'PENDIENTE',
        EnRevision: 'EN_REVISION',
        Aprobado: 'APROBADO',
        Rechazado: 'RECHAZADO'
    } as const;
    export type EstadoEnum = typeof EstadoEnum[keyof typeof EstadoEnum];
    export const EstadoSlaEnum = {
        EnTiempo: 'EN_TIEMPO',
        PorVencer: 'POR_VENCER',
        Vencido: 'VENCIDO',
        Cerrado: 'CERRADO'
    } as const;
    export type EstadoSlaEnum = typeof EstadoSlaEnum[keyof typeof EstadoSlaEnum];
}